+++ bencher/plugins/builtins.py
@@ -106,6 +106,7 @@ def _named_only_specs() -> list[tuple[str, str, Callable]]:
     from bencher.results.holoview_results.surface_result import SurfaceResult
     from bencher.results.holoview_results.table_result import TableResult
     from bencher.results.holoview_results.tabulator_result import TabulatorResult
+    from bencher.results.holoview_results.xy_curve_result import XYCurveResult
     from bencher.results.holoview_results.xy_scatter_result import XYScatterResult
     from bencher.results.rerun_result import RerunResult
     from bencher.results.video_summary import VideoSummaryResult
@@ -125,6 +126,7 @@ def _named_only_specs() -> list[tuple[str, str, Callable]]:
         ("video_summary", "panel", VideoSummaryResult.to_video_summary),
         ("rerun", "rerun", RerunResult.to_rerun),
         ("xy_scatter", "holoviews", XYScatterResult.to_plot),
+        ("xy_curve", "holoviews", XYCurveResult.to_plot),
     ]