+++ bencher/results/bench_result.py
@@ -43,6 +43,7 @@ from bencher.results.holoview_results.scatter_result import ScatterResult
 from bencher.results.holoview_results.surface_result import SurfaceResult
 from bencher.results.holoview_results.table_result import TableResult
 from bencher.results.holoview_results.tabulator_result import TabulatorResult
+from bencher.results.holoview_results.xy_curve_result import XYCurveResult
 from bencher.results.holoview_results.xy_scatter_result import XYScatterResult
 from bencher.results.optuna_result import OptunaResult
 from bencher.results.pane_result import PaneResult
@@ -77,6 +78,7 @@ class BenchResult(
     TableResult,
     TabulatorResult,
     XYScatterResult,
+    XYCurveResult,
     HoloviewResult,
     VideoSummaryResult,
     DataSetResult,